Return a "complete" transformation key.
Using a transformation key with the current FaceCropper config, allows us
to properlly use Picasso's cache.

To support this, I had to add a getSizeMode() method to FaceCropper.
The enum was already public, so a getter makes sense.

Also asked for a RGB_565 bitmap in Picasso's RequestCreator, so we don't
have to convert it later, causing extra GC.